move 'useWith' tests to separate file
