+++ types/ramda/tsconfig.json
@@ -234,6 +234,7 @@
         "test/uniq-tests.ts",
         "test/uniqWith-tests.ts",
         "test/unless-tests.ts",
-        "test/until-tests.ts"
+        "test/until-tests.ts",
+        "test/useWith-tests.ts"
     ]
 }
